Fixed time passage and added a primitive FPS counter:
I finally fixed a problem which occured almost immediately after running the program on a different computer, the particles were going way faster that what was  shown on my computer. Now, the change in locations is not dependant on performance. Also got rid of the blood-red color and added a much better nature one.